Added functionality to store github's response to DB. Also added environment files.
Created 2 components and have added service to fetch users from backend.

Added functionality to fetch users on the basis of search string.

Added a page for user details. The page shows user;s name, avatar, description, followers, following, list of his repos, etc.

Have created a separate component to dislay the search results when the user enters some string in the search bar.

Pagination applied with show-all-results, next and previous results.

Created a separate module for search for followers of a githun user. Have created a component in that module to provide the functionality to search a username.

Added another component for showing details of the follower. The module is loaded lazily and contains 2 child components, one for searching for the followers and the other to display the details of the selected follower. All working well.